rename caom2-meta-sync to darkspire
