--- caom2-meta-sync/Dockerfile
@@ -1,5 +1,0 @@
-FROM images.opencadc.org/library/cadc-java:1
-
-ADD build/distributions/caom2-meta-sync.tar /
-
-CMD ["/caom2-meta-sync/bin/caom2-meta-sync"]
