+++ darkspire/build.gradle
@@ -14,10 +14,7 @@ sourceCompatibility = 1.8
 
 group = 'org.opencadc'
 
-description = 'OpenCADC CAOM Metadata Sync application'
-def git_url = 'https://github.com/opencadc/caom2db'
-
-mainClassName = 'org.opencadc.caom2.metasync.Main'
+mainClassName = 'org.opencadc.darkspire.Main'
 
 dependencies {
     implementation 'org.opencadc:cadc-util:[1.6,2.0)'
@@ -25,11 +22,8 @@ dependencies {
     implementation 'org.opencadc:caom2persistence:[2.4.14,2.5)'
     implementation 'org.opencadc:caom2-repo:[1.4,1.5)'
 
-    // needed to run plane metadata compute plugin (--compute)
+    // needed for validation
     implementation 'org.opencadc:caom2-compute:[2.4.6,2.5)'
-
-    // needed to run access-control regen plugin (--generate-ac)
-    implementation 'org.opencadc:caom2-access-control:[2.4,2.5)'
 
     runtimeOnly 'org.postgresql:postgresql:[42.2,43.0)'
 }
